chore: remove unnecessary fluff
